Move bar chart to wait times fragment.
