+++ app/src/main/res/drawable/wait_times_marker.xml
@@ -2,7 +2,7 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:gravity="center|top">
         <shape android:shape="rectangle">
-            <corners android:radius="4dp" />
+            <corners android:radius="6dp" />
             <stroke
                 android:width="2dp"
                 android:color="@color/inactive" />
