+++ app/src/main/res/layout/activity_campus_eatery.xml
@@ -185,19 +185,11 @@
                     android:textColor="@color/primary"
                     android:textSize="18sp"/>
 
-            <RelativeLayout
-                android:id="@+id/wait_time_graph"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/wait_time_label"
-                android:background="@color/white">
-
-                <com.github.mikephil.charting.charts.BarChart
-                    android:id="@+id/wait_time_chart"
+                <FrameLayout
+                    android:id="@+id/wait_times_frame"
                     android:layout_width="match_parent"
-                    android:layout_height="156dp"/>
-
-            </RelativeLayout>
+                    android:layout_height="match_parent"
+                    android:layout_below="@id/wait_time_label" />
 
             </RelativeLayout>
