+++ app/src/main/res/layout/fragment_wait_times.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout android:id="@+id/wait_time_graph"
+    android:layout_width="fill_parent"
+    android:layout_height="wrap_content"
+    android:layout_below="@id/wait_time_label"
+    android:background="@color/white"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <com.github.mikephil.charting.charts.BarChart
+        android:id="@+id/wait_time_chart"
+        android:layout_width="match_parent"
+        android:layout_height="156dp"/>
+
+</RelativeLayout>
